Configure GPIO and toggle relay
relay is toggled using the leds library beacause, IO to a pin is IO

The big lesson here is that you MUST config the GPIO for it to work (with)
the RTOS. The builtin LEDs worked by accident.

Also, various Adafruit Feather devices do not number their pins the same across
the devices.