Get the test build to build
This does not add any tests for the Wifi bits
It only adds the fakes needed to get it to build

This is another Spike that is not TDDed or tested.  This is partly
it is using a big blob from a sample program
and partly because I am learning what it does.

More to come later.