Update kconfig so that wifi password is usable
Change the kconfig and defaults to define password in a way that can be
used in the code to detect when it is set
(this will change later when we have auto-config)

I added back some the IPV6 that was in the example. This is likely still not
really needed, but it's there so...

Changed the ERROR_CHECK uses to one that does not have NO_RETURN attributes.
This prevents the compiler from complaining about a fake that does not return.

Remove the main folder for both include and source files in the Test CMake.
There should be nothing in main/ that is not a call to compnonents/.